% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StableBiomass-package.R
\docType{package}
\name{StableBiomass-package}
\alias{StableBiomass}
\alias{StableBiomass-package}
\title{StableBiomass package}
\description{
Biomass reconstruction from stable demographic profiles.
}
\details{
The package currently provides a species-level workflow:
fertility and body mass by age are converted into stable demographic
profiles, biomass profiles, species summaries, age-class summaries, and
body-mass classes for later community analyses.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/D4v1d-d0/StableBiomass}
  \item Report bugs at \url{https://github.com/D4v1d-d0/StableBiomass/issues}
}

}
\author{
\strong{Maintainer}: David Palacios-Morales \email{davipala@ucm.es}

Authors:
\itemize{
  \item Guillermo Rodríguez-Gómez
  \item Jesús A. Martín-González
}

}
\keyword{internal}
