% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StableBiomass-package.R
\docType{package}
\name{StableBiomass-package}
\alias{StableBiomass}
\alias{StableBiomass-package}
\title{StableBiomass package}
\description{
Biomass reconstruction from stable demographic profiles.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/D4v1d-d0/StableBiomass}
  \item Report bugs at \url{https://github.com/D4v1d-d0/StableBiomass/issues}
}

}
\author{
\strong{Maintainer}: David Palacios-Morales \email{davipala@ucm.es}

Authors:
\itemize{
  \item Guillermo Rodríguez-Gómez
  \item Jesús A. Martín-González
}

}
\keyword{internal}
